added refactorings and create bucket hash
